Put back exporters to list of packages since it now can start after all pods in a cluster

+++ kubernetes/deploy.sh
@@ -7,13 +7,12 @@ echo "generate hmac secret"
 hmac=$PWD/hmac.random
 perl -e 'open(R, "< /dev/urandom") or die; sysread(R, $K, 20) or die; print $K' > $hmac
 
-pkgs="das dbs ing frontend couchdb reqmgr httpsgo reqmon workqueue tfaas"
+pkgs="das dbs ing frontend couchdb reqmgr httpsgo reqmon workqueue tfaas exporters"
 
 echo "### secrets"
 for p in $pkgs; do
     kubectl delete secret/${p}-secrets
 done
-kubectl delete secret/exporters-secrets
 kubectl delete secret/dbs2go-secrets
 kubectl -n kube-system delete secret/traefik-cert
 kubectl -n kube-system delete configmap traefik-conf
@@ -49,7 +48,6 @@ echo "### apply secrets"
 for p in $pkgs; do
     kubectl apply -f ${p}-secrets.yaml --validate=false
 done
-kubectl apply -f exporters-secrets.yaml --validate=false
 kubectl apply -f dbs2go-secrets.yaml --validate=false
 rm *secrets.yaml $hmac
 
@@ -72,7 +70,6 @@ for p in $pkgs; do
         kubectl delete -f ${p}.yaml
     fi
 done
-kubectl delete -f exporters.yaml
 kubectl delete -f httpgo.yaml
 kubectl delete -f das2go.yaml
 kubectl delete -f dbs2go.yaml
@@ -93,7 +90,9 @@ kubectl apply -f dbs2go.yaml --validate=false
 sleep 2
 echo
 echo "### delete daemon ingress-traefik"
-kubectl delete daemonset ingress-traefik -n kube-system
+if [ -z "`kubectl get daemonset -n kube-system | grep ingress-traefik`" ]; then
+    kubectl delete daemonset ingress-traefik -n kube-system
+fi
 sleep 2
 echo "### deploy traefik"
 kubectl -n kube-system create secret generic traefik-cert --from-file=$server_crt --from-file=$server_key
@@ -107,7 +106,9 @@ kubectl -n monitoring delete -f kubernetes-prometheus/prometheus-service.yaml
 kubectl -n monitoring delete -f kubernetes-prometheus/prometheus-deployment.yaml
 kubectl -n monitoring delete -f kubernetes-prometheus/config-map.yaml
 
-kubectl create namespace monitoring
+if [ -z "`kubectl get namespaces | grep monitoring`" ]; then
+    kubectl create namespace monitoring
+fi
 kubectl -n monitoring apply -f kubernetes-prometheus/config-map.yaml --validate=false
 kubectl -n monitoring apply -f kubernetes-prometheus/prometheus-deployment.yaml --validate=false
 kubectl -n monitoring apply -f kubernetes-prometheus/prometheus-service.yaml --validate=false
@@ -118,21 +119,3 @@ echo "### we may access prometheus locally as following"
 echo "kubectl -n monitoring port-forward $prom 8080:9090"
 echo "### to access prometheus externally we should do the following:"
 echo "ssh -S none -L 30000:$kubehost:30000 $USER@lxplus.cern.ch"
-
-echo
-echo "Wait until pods will be running and deploy exporters"
-sleep 60
-
-# we can deploy explorers.yaml once all services which it monitors are started
-for i in 1 2 3 4 5; do
-    notRun=`kubectl get pods | grep -v Running | grep -v READY`
-    if [ -n "$notRun" ]; then
-        echo "attempt: $i"
-        echo "Services that are not run yet:"
-        echo $notRun
-        sleep 60
-        continue
-    fi
-    kubectl apply -f exporters.yaml --validate=false
-    break
-done
